modification logique modification d'un produit

+++ backend/controllers/sauces.js
@@ -19,18 +19,26 @@ exports.createSauce = (req, res, next) => {
 exports.modifySauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
-      const filename = sauce.imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
-        const sauceObject = req.file
-          ? {
-              ...JSON.parse(req.body.sauce),
-              imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
-            }
-          : { ...req.body };
-        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+      if (req.file) {
+        // si la requête comporte une modification de l'image
+        // on supprime l'image précédente de la base
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          // on crée ensuite un nouvel objet avec la requête et l'image
+          const sauceObject = {
+            ...JSON.parse(req.body.sauce),
+            imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
+          };
+          Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
+            .catch((error) => res.status(400).json({ error }));
+        });
+      } else {
+        // si la requête ne comporte pas de modification de l'image, on renvoit la requête updateOne
+        Sauce.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
           .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
           .catch((error) => res.status(400).json({ error }));
-      });
+      }
     })
     .catch((error) => res.status(500).json({ error }));
 };
